Extract AlphaBeta search into dedicated class
